requirements.txt
streamlit==1.40.0
pandas==2.2.3
python-dateutil==2.9.0.post0
numpy==1.26.4
